[FIX]购物车对象校验修复(针对购物车拆分数量的场景)

See merge request operation-srm-watson/srm-mall-watsons!25

+++ src/main/java/org/srm/mall/other/app/service/impl/WatsonsShoppingCartServiceImpl.java
@@ -275,9 +275,18 @@ public class WatsonsShoppingCartServiceImpl extends ShoppingCartServiceImpl impl
                     productStockService.productStockConsumption(null, shoppingCart.getProductId(), shoppingCart.getQuantity().longValue(), true);
                 }
                 updateBudgetInfoResult(shoppingCart, budgetSwitch);
-                shoppingCartRepository.deleteByPrimaryKey(shoppingCart.getCartId());
+                List<ShoppingCart> shoppingCarts = shoppingCartRepository.selectByIds(shoppingCart.getCartId().toString());
+                if(CollectionUtils.isNotEmpty(shoppingCarts)){
+                    ShoppingCart currentShoppingCart = shoppingCarts.get(0);
+                    if(currentShoppingCart.getQuantity()-shoppingCart.getQuantity()<=0L){
+                        //删除购物车对象
+                        shoppingCartRepository.deleteByPrimaryKey(shoppingCart.getCartId());
+                    }else {
+                        currentShoppingCart.setQuantity(currentShoppingCart.getQuantity()-shoppingCart.getQuantity());
+                        shoppingCartRepository.updateOptional(currentShoppingCart,ShoppingCart.FIELD_QUANTITY);
+                    }
+                }
             }
-
             //遍历每个可以提交的预采申请订单结束
         }
         PurchaseRequestVO result;
